remove asset from assets fix with user selections

+++ src/app/configurations/typical-configurations/typical-configurations.page.ts
@@ -1,9 +1,12 @@
 import { AfterViewInit, ChangeDetectorRef, Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { IonSearchbar, NavController } from '@ionic/angular';
-import { debounceTime, Subscription } from 'rxjs';
+import { AlertController } from '@ionic/angular';
+import { debounceTime, EMPTY, from, map, Subscription, tap } from 'rxjs';
+import {AlertConfirmationType} from '../../shared/models/alert-confirmation.enum';
 import { Asset } from '../models/asset.model';
 import { AssetsService } from '../services/assets/assets.service';
+import { UserSelectionService } from 'src/app/shared/services/user-selection.service';
 
 @Component({
   selector: 'app-typical-configurations',
@@ -16,13 +19,17 @@ export class TypicalConfigurationsPage implements OnInit, AfterViewInit, OnDestr
   areAssetsFound: boolean = true;
   assetsSuscription!: Subscription;
 
+  alertSubscription!: Subscription;
+
 
 
   constructor(
     private assetsService:AssetsService,
     private cd: ChangeDetectorRef,
     private route: Router,
-    private navCntrl: NavController) { }
+    private navCntrl: NavController,
+    private userSelectionService: UserSelectionService,
+    private alertController: AlertController) { }
 
 
 
@@ -36,15 +43,24 @@ export class TypicalConfigurationsPage implements OnInit, AfterViewInit, OnDestr
 
     ngOnDestroy(): void {
       this.assetsSuscription.unsubscribe();
+      this.alertSubscription.unsubscribe();
     }
 
     onAssetLink(assetId: string){
     this.route.navigate(['configurations', 'typical-configurations', assetId]);
     }
 
-    onRemoveAsset(assetId:string){
-      this.assetsService.removeAsset(assetId);
+    onRemoveAsset(assetId:string, $event: Event){
+      $event.stopPropagation();
 
+      let isAssetInUserSelection = this.userSelectionService.getIsAssetInAssetsForPDF(assetId);
+      if(isAssetInUserSelection){
+          this.alertSubscription = from(this.presentAlert()).subscribe(()=>{
+            this.userSelectionService.setIsUserSelectionsMenuOpen(true);
+          });
+      }else{
+        this.assetsService.removeAsset(assetId);
+      }
     }
 
     ngAfterViewInit(): void {
@@ -56,6 +72,24 @@ export class TypicalConfigurationsPage implements OnInit, AfterViewInit, OnDestr
         this.areAssetsFound = this.assetsList.length === 0? false : true;
       })
 
+    }
+
+    async presentAlert() {
+      const alert = await this.alertController.create({
+        header: 'Please remove the asset from your selections first.',
+        cssClass: 'custom-alert',
+        buttons: [
+          {
+            text: 'OK',
+            role: AlertConfirmationType.Confirm,
+          },
+        ],
+      });
+
+      await alert.present();
+
+      const { role } = await alert.onDidDismiss();
+      return role as AlertConfirmationType;
     }
